==> 1.series.ipynb <==
import pandas as pd

dados = [10, 20, 30, 40, 50]
serie = pd.Series(dados)
serie

type (serie)

vendas_mensais = pd.Series(
    [200, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500, 7000],
    index=[
        'jan', 'fev', 'mar', 'abr', 'mai', 'jun', 'jul', 'ago', 'set', 'out', 'nov']
)

vendas_mensais

#Calcular  e exibir a média de vendas
media_vendas = vendas_mensais.mean()
media_vendas

#Identificando o mês com maior venda
mes_maior_venda = vendas_mensais.idxmax()
mes_maior_venda

#Exibir as vendas acima da media
vendas_acima_media = vendas_mensais[vendas_mensais > media_vendas]
vendas_acima_media
