==> 2.dataframe.ipynb <==
import pandas as pd

dados = {
    'nome': ['Ana', 'Bia', 'Carlos', 'Daniel'],
    'idade': [23, 34, 45, 56],
    'cidade': ['São Paulo', 'Rio de Janeiro', 'Belo Horizonte', 'Curitiba']
}

df = pd.DataFrame(dados)
# Criando um DataFrame a partir de um dicionário

df # Exibindo o DataFrame df é variavel que armazena o DataFrame

df.head(1) # Exibindo as 5 primeiras linhas do DataFrame,
 # () sem argumentos, o método head() exibe as 5 primeiras linhas do DataFrame	

df.tail(1) # Exibindo as 2 últimas linhas do DataFrame
# tail é o oposto de head, ou seja, exibe as últimas linhas do DataFrame

dados_notas ={
    'Nome': ['Ana', 'Bia', 'Carlos', 'Daniel'],
    'Matematica': [9, 8, 7, 6],
    'Fisica': [7, 8, 9, 6],
    'Quimica': [8, 9, 7, 6]
}

df_notas = pd.DataFrame(dados_notas)

df_notas

#Calculando a media das notas
df_notas

df_notas['Média'] = (df_notas['Matematica'] + df_notas['Fisica'] + df_notas['Quimica']) / 3
#Escreva CERTO o nome das colunas, se não, não puxa
df_notas

#Identificar alunos aprovados com nota boa (precisa ser maior media >=82.5)
df_notas_aprovados = df_notas[df_notas['Média'] >= 8]
df_notas_aprovados


df_notas [['Nome', 'Média']]

df_notas_aprovados [['Nome','Média']]
