==> 5.importa.csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("dados/dados.csv", sep=";")
df_produtos

df_produtos.head(8)#Primeiras 8 linhas

df_produtos.tail(8) #Ultimos 8 registros

df_produtos.shape#(linhas, colunas)

df_produtos.dtypes#Verifica os tipos de dados

df_produtos.sort_values(by='Quantidade', ascending=False) #Ordena pela quantidade de produtos em ordem decrescente

df_produtos['Categoria'].values #Verifica os valores únicos da coluna Categoria
