==> 6.analise_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("dados/dados.csv", sep=';')
df_produtos

df_produtos.dtypes

df_produtos ['Quantidade'] = pd.to_numeric(df_produtos['Quantidade'])
df_produtos['Preço'] = pd.to_numeric(df_produtos['Preço'])

#criar uma cópia do dataframe
vendas_categoria = df_produtos.groupby(['Categoria'])[['Quantidade', 'Preço']].sum()
vendas_categoria

# Qual foi o produto mais vendido?
produto_mais_vendido = df_produtos.groupby('Produto')['Quantidade'].sum().idxmax()
#idmax() retorna o índice do maior valor
produto_mais_vendido

#Total de unidades vendidas por produto
produto_mais_vendido = df_produtos.groupby(['Produto'])[['Quantidade']].sum()
produto_mais_vendido

#Preço médio por produto
media_preco_produto = df_produtos.groupby(['Produto'])[['Preço']].mean()
media_preco_produto

#Correlação entre Unidades Vendidas e Preço Unitário
correlacao = df_produtos['Quantidade'].corr(df_produtos['Preço'])
correlacao
# Quer dizer que quanto mais unidades vendidas, maior o preço unitário? Não necessariamente. A correlação não implica causalidade.
# Pode ser que o preço unitário seja maior para produtos que são vendidos em menor quantidade.

vendas_categoria = df_produtos.groupby('Categoria')['Quantidade'].sum()
vendas_categoria_df = vendas_categoria.reset_index()
vendas_categoria_df.columns = ['Categoria', 'Unidades Vendidas']


vendas_categoria_df

vendas_categoria_df.to_csv('dados/vendas_categoria.csv', sep=';', index=False)
# Exportando o dataframe para um arquivo CSV
