==> 7.criarplanilha.ipynb <==
import pandas as pd
import openpyxl

dados = {
    'Time': ['Flamengo', 'Fluminense', 'Vasco', 'Botafogo', 'Corinthians'],
    'Partidas_Jogadas': [10, 10, 10, 10, 10],
    'Vitorias': [5, 3, 2, 4, 6],
    'Empates': [3, 4, 5, 2, 2],
    'Derrotas': [2, 3, 3, 4, 2],
    'Gols_Pro': [20, 15, 10, 12, 25],
    'Gols_Contra': [10, 12, 15, 10, 8],
}

df = pd.DataFrame(dados)
df

df['Saldo_Gols'] = df['Gols_Pro'] - df['Gols_Contra']
df

df['Pontos'] = (df['Vitorias'] * 3) + df['Empates'] * 1
df

df = df.sort_values(by='Pontos', ascending=False)
df

df.to_excel('classificacao_campeonato.xlsx', index=False)
