==> 8.analise_planilha.ipynb <==
import pandas as pd

df = pd.read_excel("dados/classificacao_campeonato.xlsx")
df

df.describe()

df.head(2)

df.tail(2)

#Analise de desempenho por time
desempenho_info = df[["Time", "Pontos", "Vitorias", "Saldo_Gols"]].set_index("Time")
desempenho_info

#Time com mais vitorias e menos vitorias
time_mais_vitorias = df[df["Vitorias"] == df["Vitorias"].max()]
time_mais_vitorias

#Time com mais vitorias e menos vitorias
time_mais_vitorias = df[df["Vitorias"] == df["Vitorias"].min()]
time_mais_vitorias

# Distribuição de Gols pró
df["Gols_Pro"].value_counts()

#Relação entre Gols Pro e pontos
correlacao = df["Gols_Pro"].corr(df['Pontos'])
correlacao

# Distribuição de Gols contra
correlacao = df["Gols_Contra"].corr(df['Pontos'])
correlacao
