==> 12.bd.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

engine = create_engine('sqlite:///ecommerce.db')

df = pd.read_csv("dados/dados.csv", sep=";")
df

df.head()

try:
    df.to_sql('dados', con=engine, if_exists='replace', index=False)
    print("Dados inseridos com sucesso!")
except Exception as e:
    print(f"Ocorreu um erro ao inserir os dados: {e}")
